feat: pipeline false alarm checks, 64-bit chain counts, Markov NTLM8/9 tests, improved coverage model
- Pipeline false alarm checking in crackalack_lookup (launch/harvest pattern for GPU overlap)
- Widen chain count fields to uint64_t to support large tables (>4B chains)
- Add per-kernel GWS tuning (get_optimal_gws now takes kernel name)
- Improve coverage model in crackalack_plan with iterative merge accounting
- Add Markov NTLM8/NTLM9 chain test files and unit test registrations
- Update test_markov_accuracy.sh with train/test split and improved params
- Fix Metal kernel address space qualifiers and buffer argument handling
- Various kernel fixes for Markov fast-path (pspace_table, sorted_bigram)

+++ rtc_decompress.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-int rtc_decompress(char *filename, uint64_t **uncompressed_table, unsigned int *num_chains);
+int rtc_decompress(char *filename, uint64_t **uncompressed_table, uint64_t *num_chains);
 
 #endif
